requirements.txt
pyaudio
openai-whisper
queue
numpy
lmstudio
pyttsx3
pyqt5